feat: Update shader for Minecraft 1.21.5

+++ src/main/java/cn/pupperclient/utils/mouse/ScrollHelper.java
@@ -13,7 +13,7 @@ public class ScrollHelper {
 	}
 
 	public void onScroll(double amount) {
-		scroll += amount * 60;
+		scroll += (float) (amount * 60);
 	}
 
 	public void onUpdate() {
